+++ scenes-plugin/src/main/java/gg/bundlegroup/scenes/command/tag/entity/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package gg.bundlegroup.scenes.command.tag.entity;
+
+import org.jspecify.annotations.NullMarked;