+++ scenes-plugin/src/main/java/gg/bundlegroup/scenes/command/tag/worldedit/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package gg.bundlegroup.scenes.command.tag.worldedit;
+
+import org.jspecify.annotations.NullMarked;