+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-Initial release.
+* Add WorldEdit selection commands
+    * Allows adding/removing tags for all entities in your WorldEdit selection
+    * `/scenes worldedit tag add ...`
+    * `/scenes worldedit tag remove ...`